# move benchmarks to separate package, speed up CI builds