--- a/.travis/script.sh
+++ b/.travis/script.sh
@@ -4,16 +4,16 @@ set -e
 
 go get -t ./...
 if [ ${TESTMODE} == "unit" ]; then
-  ginkgo -r --cover --randomizeAllSpecs --randomizeSuites --trace --progress --skipPackage integrationtests --skipMeasurements
+  ginkgo -r --cover --randomizeAllSpecs --randomizeSuites --trace --progress --skipPackage integrationtests,benchmark
 fi
 
 if [ ${TESTMODE} == "integration" ]; then
   # run benchmark tests
-  ginkgo --randomizeAllSpecs --randomizeSuites --trace --progress -focus "Benchmark"
+  ginkgo --randomizeAllSpecs --randomizeSuites --trace --progress benchmark
   # run benchmark tests with the Go race detector
   # The Go race detector only works on amd64.
   if [ ${TRAVIS_GOARCH} == 'amd64' ]; then
-    ginkgo --race --randomizeAllSpecs --randomizeSuites --trace --progress -focus "Benchmark"
+    ginkgo --race --randomizeAllSpecs --randomizeSuites --trace --progress benchmark
   fi
   # run integration tests
   ginkgo -v -r --randomizeAllSpecs --randomizeSuites --trace --progress integrationtests