--- a/.travis/script.sh
+++ b/.travis/script.sh
@@ -9,11 +9,11 @@ fi
 
 if [ ${TESTMODE} == "integration" ]; then
   # run benchmark tests
-  ginkgo --randomizeAllSpecs --randomizeSuites --trace --progress benchmark
+  ginkgo --randomizeAllSpecs --randomizeSuites --trace --progress benchmark -- -samples=1
   # run benchmark tests with the Go race detector
   # The Go race detector only works on amd64.
   if [ ${TRAVIS_GOARCH} == 'amd64' ]; then
-    ginkgo --race --randomizeAllSpecs --randomizeSuites --trace --progress benchmark
+    ginkgo --race --randomizeAllSpecs --randomizeSuites --trace --progress benchmark -- -samples=1 -size=10
   fi
   # run integration tests
   ginkgo -v -r --randomizeAllSpecs --randomizeSuites --trace --progress integrationtests