--- a/benchmark/benchmark_suite_test.go
+++ b/benchmark/benchmark_suite_test.go
@@ -1,0 +1,13 @@
+package benchmark
+
+import (
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+
+	"testing"
+)
+
+func TestBenchmark(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Benchmark Suite")
+}