Compared to #764, which already reduced the Travis build time by 100s, this PR achieves a further speedup by reducing the number of samples and the data size transferred (for the test run with the race detector), thus saving another:
* 25s on the benchmark tests run without the race detector (from 50s down to 25s)
* 160s on the benchmark tests run without the race detector (from ~170s down to less than 10s)

In total, we can reduce the Travis build time by around 3 minutes by merging this PR.
On AppVeyor, we save about 10 - 15s on every build, which is nice, however, that was never our bottleneck.